src/main/java/ru/netology/manager/MovieManager.java
package ru.netology.manager;

import ru.netology.domain.Movie;

public class MovieManager {
    private Movie[] movies = new Movie[0];
    int movieLimit = 10;

    public MovieManager(int movieCount) {
        this.movieLimit = movieCount;
    }

    public  MovieManager() {}

    public void add(Movie item) {

        int length = movies.length + 1;
        Movie[] tmp = new Movie[length];

        System.arraycopy(movies, 0, tmp, 0, movies.length);
        int lastIndex = tmp.length - 1;
        tmp[lastIndex] = item;
        movies = tmp;
    }

    public Movie[] getAll() {
        Movie[] result = new Movie[movies.length];

        for (int i = 0; i < result.length; i++) {
            int index = movies.length - i - 1;
            result[i] = movies[index];
        }
        return result;
    }

    public Movie[] getMovie() {
        int count = movieLimit;
        if (movies.length < count) {
            count = movies.length;
        }

        Movie[] result = new Movie[count];
        for (int i = 0; i < result.length; i++) {
            int index = movies.length - i - 1;
            result[i] = movies[index];
        }
        return result;
    }


}
